Implement templates

+++ lib/templates.rb
@@ -7,72 +7,57 @@ require "uri"
 require "versions"
 
 module SendgridTemplateEngine
-  # Your code goes here...
+
   class Templates
 
-    # TODO 関数にコードブロックを渡すにはどうすればよいか？
-    def get_all(username, password)
-
-      endpoint = "https://#{username}:#{password}@api.sendgrid.com/v3/templates"
-      body = RestClient.get(endpoint).body
-      response = JSON.parse(body)
-
-      if response["errors"] then
-        raise response["errors"][0]
-      else
-        temps = []
-        response["templates"].each{|template|
-          temp = Template.create(template)
-          temps.push(temp)
-        } if response["templates"] != nil
-        temps
-      end
+    def initialize(username, password)
+      raise ArgumentError.new("username should not be nil") if username == nil
+      raise ArgumentError.new("password should not be nil") if password == nil
+      @username = username
+      @password = password
     end
 
-    def get(username, password, template_id)
-
-      # TODO template_id nilチェック
-      endpoint = "https://#{username}:#{password}@api.sendgrid.com/v3/templates/#{template_id}"
+    def get_all()
+      endpoint = "https://#{@username}:#{@password}@api.sendgrid.com/v3/templates"
       body = RestClient.get(endpoint).body
       response = JSON.parse(body)
-
-      if response["errors"] then
-        raise response["errors"][0]
-      else
-        Template.create(response)
-      end
+      temps = []
+      response["templates"].each{|template|
+        temp = Template.create(template)
+        temps.push(temp)
+      } if response["templates"] != nil
+      temps
     end
 
-    def post(username, password, name)
-      endpoint = "https://#{username}:#{password}@api.sendgrid.com/v3/templates"
+    def get(template_id)
+      raise ArgumentError.new("template_id should not be nil") if template_id == nil
+      endpoint = "https://#{@username}:#{@password}@api.sendgrid.com/v3/templates/#{template_id}"
+      body = RestClient.get(endpoint).body
+      Template.create(JSON.parse(body))
+    end
+
+    def post(name)
+      raise ArgumentError.new("name should not be nil") if name == nil
+      endpoint = "https://#{@username}:#{@password}@api.sendgrid.com/v3/templates"
       params = Hash.new
       params['name'] = name
-      b = RestClient.post endpoint, params.to_json, :content_type => :json
-      response = JSON.parse(b.body)
-
-      if response["errors"] then    # TODO 正常応答の場合、errorsは来ないと思われる。例外の時に処理する要修正
-        raise response["errors"][0]
-      else
-        Template.create(response)
-      end
+      body = RestClient.post(endpoint, params.to_json, :content_type => :json).body
+      Template.create(JSON.parse(body))
     end
 
-    def patch(username, password, template_id, name)
-      endpoint = "https://#{username}:#{password}@api.sendgrid.com/v3/templates/#{template_id}"
+    def patch(template_id, name)
+      raise ArgumentError.new("template_id should not be nil") if template_id == nil
+      raise ArgumentError.new("name should not be nil") if name == nil
+      endpoint = "https://#{@username}:#{@password}@api.sendgrid.com/v3/templates/#{template_id}"
       params = Hash.new
       params['name'] = name
       body = RestClient.patch(endpoint, params.to_json, :content_type => :json).body
-      response = JSON.parse(body)
-
-      if response["errors"] then
-        raise response["errors"][0]
-      else
-        Template.create(response)
-      end
+      Template.create(JSON.parse(body))
     end
 
-    def delete(username, password, template_id)
-      endpoint = "https://#{username}:#{password}@api.sendgrid.com/v3/templates/#{template_id}"
+    def delete(template_id)
+      raise ArgumentError.new("template_id should not be nil") if template_id == nil
+      endpoint = "https://#{@username}:#{@password}@api.sendgrid.com/v3/templates/#{template_id}"
       RestClient.delete(endpoint)
     end
